Show google login popup on user click
Refactor google auth flow, to make the google
login popup be iniated right after the user click
(in the login button click handler),
in order to avoid popup blocking by browser.
See this article: http://arveknudsen.com/?p=368

Previously, on the login button click, the app
was loading google client, and only then initating the popup.
Now, the google client is loaded before hand,
and when user clicks the login button, google login
is called at once.
This way there is no intermediary operations between
the user click and showing the popup,
making the browser to relax on blocking it.