Fix a bug with consequent note name updates
The function to update note name was relying on the supplied note name
to find the note content file, resulting in problems, as the client
doesn't know what is the current name.

Now when updating note name, first fetch the note by id, to get its name
as on Google Drive, and use that name to find the note content file.

So now only note id is used when updating the note name.

Fix #110