Remove unused 'children' and 'parent' group tag
When VisNetwork library was used, it allowed adding group tags to nodes,
and a different color was used for groups 'children' and 'parent'.
The 'parent' group was for nodes which have children, and the 'children'
group was for nodes without children.
This concept was quite useless, so it is not used after introducing a
custom mind map rendering engine.