Make change parent work after page refresh
Since the change parent page has its own route, user can refresh the
page and in this case "isChangeParentModeActive" wasn't set to true, as
it was set on "change parent" button click.
Now set this variable to true on change parent page load as well.

Not sure if this is the right approach UX-wise, or maybe it's better to
simply redirect the user back to the mind map in this case, but it's
handy for developing the change parent page as I don't need to click the
"change parent" button every time the page refreshes :)