Handle fetching a trashed note
Since notes mind map is saved in local storage, a selected note may turn
out to be already trashed (if user deleted in on another device).
In this case show a dialog explaining the user what happened, and when
the dialog is dismissed, change selected note to the root note.